refactor: consolidate resume logic under waitpoint folder
Extract resumeFromWaitpoint, handleSyncResumeFlow, and enqueueResume
from flow-run-service into waitpoint/resume-service. Move waitpoint
resume routes into waitpoint/waitpoint-resume-controller and relocate
legacy-resume-controller under waitpoint/ alongside the new controller.
