+++ packages/server/api/src/app/flows/flow-run/flow-run-controller.ts
@@ -1,7 +1,5 @@
 import {
     ActivepiecesError,
-    ALL_PRINCIPAL_TYPES,
-    apId,
     ApId,
     BulkActionOnRunsRequestBody,
     BulkArchiveActionOnRunsRequestBody,
@@ -17,7 +15,6 @@ import {
     SeekPage,
     SERVICE_KEY_SECURITY_OPENAPI,
 } from '@activepieces/shared'
-import { FastifyBaseLogger, FastifyReply } from 'fastify'
 import { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import { StatusCodes } from 'http-status-codes'
 import { z } from 'zod'
@@ -25,7 +22,6 @@ import { ProjectResourceType } from '../../core/security/authorization/common'
 import { securityAccess } from '../../core/security/authorization/fastify-security'
 import { FlowRunEntity } from './flow-run-entity'
 import { flowRunService } from './flow-run-service'
-import { waitpointService } from './waitpoint/waitpoint-service'
 
 const DEFAULT_PAGING_LIMIT = 10
 
@@ -58,32 +54,6 @@ export const flowRunController: FastifyPluginAsyncZod = async (app) => {
             await reply.send(flowRun)
         },
     )
-
-    app.all('/:id/waitpoints/:waitpointId', ResumeByWaitpointRequest, async (req, reply) => {
-        const headers = req.headers as Record<string, string>
-        const queryParams = req.query as Record<string, string>
-        await handleAsyncResume({ flowRunId: req.params.id, waitpointId: req.params.waitpointId, body: req.body, headers, queryParams, log: req.log, reply })
-    })
-
-    app.all('/:id/waitpoints/:waitpointId/sync', ResumeByWaitpointRequest, async (req, reply) => {
-        const headers = req.headers as Record<string, string>
-        const queryParams = req.query as Record<string, string>
-        await handleSyncResume({ flowRunId: req.params.id, waitpointId: req.params.waitpointId, body: req.body, headers, queryParams, log: req.log, reply, correlationId: req.params.waitpointId })
-    })
-
-    app.all('/:id/requests/:requestId', ResumeFlowRunRequest, async (req, reply) => {
-        const waitpoint = await waitpointService(req.log).getByFlowRunId(req.params.id)
-        const headers = req.headers as Record<string, string>
-        const queryParams = req.query as Record<string, string>
-        await handleAsyncResume({ flowRunId: req.params.id, waitpointId: waitpoint?.id ?? apId(), body: req.body, headers, queryParams, log: req.log, reply })
-    })
-
-    app.all('/:id/requests/:requestId/sync', ResumeFlowRunRequest, async (req, reply) => {
-        const waitpoint = await waitpointService(req.log).getByFlowRunId(req.params.id)
-        const headers = req.headers as Record<string, string>
-        const queryParams = req.query as Record<string, string>
-        await handleSyncResume({ flowRunId: req.params.id, waitpointId: waitpoint?.id ?? apId(), body: req.body, headers, queryParams, log: req.log, reply, correlationId: req.params.requestId })
-    })
 
     app.post('/:id/retry', RetryFlowRequest, async (req) => {
         const flowRun = await flowRunService(req.log).retry({
@@ -190,30 +160,6 @@ const GetRequest = {
     },
 }
 
-const ResumeByWaitpointRequest = {
-    config: {
-        security: securityAccess.unscoped(ALL_PRINCIPAL_TYPES),
-    },
-    schema: {
-        params: z.object({
-            id: ApId,
-            waitpointId: z.string(),
-        }),
-    },
-}
-
-const ResumeFlowRunRequest = {
-    config: {
-        security: securityAccess.unscoped(ALL_PRINCIPAL_TYPES),
-    },
-    schema: {
-        params: z.object({
-            id: ApId,
-            requestId: z.string(),
-        }),
-    },
-}
-
 const RetryFlowRequest = {
     config: {
         security: securityAccess.project(
@@ -273,43 +219,4 @@ const BulkRetryFlowRequest = {
     },
 }
 
-async function handleAsyncResume({ flowRunId, waitpointId, body, headers, queryParams, log, reply }: ResumeHandlerParams): Promise<void> {
-    const { stale } = await flowRunService(log).resumeFromWaitpoint({
-        flowRunId,
-        waitpointId,
-        resumePayload: { body, headers, queryParams },
-    })
-    if (stale) {
-        await reply.send({
-            message: 'This link has expired. The action may have already been processed.',
-        })
-        return
-    }
-    await reply.send({
-        message: 'Your response has been recorded. You can close this page now.',
-    })
-}
 
-async function handleSyncResume({ flowRunId, waitpointId, body, headers, queryParams, log, reply, correlationId }: ResumeHandlerParams & { correlationId: string }): Promise<void> {
-    const response = await flowRunService(log).handleSyncResumeFlow({
-        runId: flowRunId,
-        waitpointId,
-        payload: {
-            body,
-            headers,
-            queryParams,
-        },
-        correlationId,
-    })
-    await reply.status(response.status).headers(response.headers).send(response.body)
-}
-
-type ResumeHandlerParams = {
-    flowRunId: string
-    waitpointId: string
-    body: unknown
-    headers: Record<string, string>
-    queryParams: Record<string, string>
-    log: FastifyBaseLogger
-    reply: FastifyReply
-}
